+++ tests/testthat/test-calculate_source_wads.R
@@ -4,5 +4,5 @@ test_that("works correctly", {
 
 test_that("wrong input type fails", {
   expect_error(calculate_source_wads(example_source_object),
-               "sample_data must be of class <qsip_sample_data>, not qsip_source_data")
+               "sample_data should be of class <qsip_sample_data>")
 })