+++ tests/testthat/test-calculate_tube_rel_abundance.R
@@ -9,13 +9,13 @@ test_that("make sure data types are correct", {
   expect_error(calculate_tube_rel_abundance(example_source_df,
                                             example_sample_object,
                                             example_feature_object),
-               "source_data must be of class <qsip_source_data>")
+               "source_data should be of class <qsip_source_data>")
   expect_error(calculate_tube_rel_abundance(example_source_object,
                                             example_sample_df,
                                             example_feature_object),
-               "sample_data must be of class <qsip_sample_data>")
+               "sample_data should be of class <qsip_sample_data>")
   expect_error(calculate_tube_rel_abundance(example_source_object,
                                             example_sample_object,
                                             example_feature_df),
-               "feature_data must be of class <qsip_feature_data>")
+               "feature_data should be of class <qsip_feature_data>")
 })