+++ tests/testthat/test-qsip_sample_data.R
@@ -1,7 +1,7 @@
-add_taxonomy_testdf <- readRDS(test_path("fixtures", "sample_data_rel_amt.rds"))
+sample_data_rel_amt <- readRDS(test_path("fixtures", "sample_data_rel_amt.rds"))
 
 test_that("works correctly", {
-  expect_snapshot(qsip_sample_data(add_taxonomy_testdf,
+  expect_snapshot(qsip_sample_data(sample_data_rel_amt,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -32,7 +32,7 @@ test_that("no gradient_pos_rel_amt given will autocalculate", {
 })
 
 test_that("missing columns give errors", {
-  expect_error(qsip_sample_data(add_taxonomy_testdf,
+  expect_error(qsip_sample_data(sample_data_rel_amt,
     sample_id = "not_found",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -40,7 +40,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "sample_id column not found")
-  expect_error(qsip_sample_data(add_taxonomy_testdf,
+  expect_error(qsip_sample_data(sample_data_rel_amt,
     sample_id = "sample",
     source_mat_id = "not_found",
     gradient_position = "Fraction",
@@ -48,7 +48,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "source_mat_id column not found")
-  expect_error(qsip_sample_data(add_taxonomy_testdf,
+  expect_error(qsip_sample_data(sample_data_rel_amt,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "not_found",
@@ -56,7 +56,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "gradient_position column not found")
-  expect_error(qsip_sample_data(add_taxonomy_testdf,
+  expect_error(qsip_sample_data(sample_data_rel_amt,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -64,7 +64,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "gradient_pos_density column not found")
-  expect_error(qsip_sample_data(add_taxonomy_testdf,
+  expect_error(qsip_sample_data(sample_data_rel_amt,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -72,7 +72,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "not_found",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "gradient_pos_amt column not found")
-  expect_error(qsip_sample_data(add_taxonomy_testdf,
+  expect_error(qsip_sample_data(sample_data_rel_amt,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -80,4 +80,16 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "not_found"
   ), "gradient_pos_rel_amt column not found")
+})
+
+
+test_that("duplicate sample_ids give error", {
+  expect_error(qsip_sample_data(rbind(sample_data_rel_amt,sample_data_rel_amt),
+                                   sample_id = "sample",
+                                   source_mat_id = "source",
+                                   gradient_position = "Fraction",
+                                   gradient_pos_density = "density_g_ml",
+                                   gradient_pos_amt = "avg_16S_g_soil",
+                                   gradient_pos_rel_amt = "gradient_pos_rel_amt"
+  ), "Some sample_ids are duplicated")
 })