+++ tests/testthat/test-summarize_EAF_values.R
@@ -1,19 +1,10 @@
-# make complete qSIP object
-test_qsip <- qsip_data(example_source_object, example_sample_object, example_feature_object) |>
-  run_feature_filter(
-    unlabeled_source_mat_ids = get_all_by_isotope(example_qsip_object, "12C"),
-    labeled_source_mat_ids = c("S178", "S179", "S180"),
-    min_unlabeled_sources = 3,
-    min_labeled_sources = 3,
-    min_unlabeled_fractions = 6,
-    min_labeled_fractions = 6
-  ) |>
-  run_resampling(resamples = 1000, with_seed = 43) |>
-  run_EAF_calculations()
+normal_qsip <- readRDS(test_path("fixtures", "qsip_normal_strict_EAF.rds"))
+normal_qsip_failures <- readRDS(test_path("fixtures", "qsip_normal_failures_EAF.rds"))
+multi_qsip <- readRDS(test_path("fixtures", "multi_qsip_EAF.rds"))
 
 test_that("snapshots look as expected", {
-  expect_snapshot(summarize_EAF_values(test_qsip))
-  expect_snapshot(summarize_EAF_values(test_qsip, confidence = 0.95))
+  expect_snapshot(summarize_EAF_values(normal_qsip))
+  expect_snapshot(summarize_EAF_values(normal_qsip, confidence = 0.95))
 })
 
 test_that("Wrong input types give error", {
@@ -38,14 +29,13 @@ test_that("Wrong input types give error", {
 test_that("Make sure qsip object has EAF data", {
   expect_error(
     summarize_EAF_values(example_qsip_object),
-    "ERROR: @EAF slot is empty, have you run run_EAF_calculations()?"
+    "object is a non-filtered <qsip_data> object"
   )
 })
 
 
-
 test_that("works on lists of qsip objects", {
-  expect_snapshot(summarize_EAF_values(list("A" = test_qsip)))
+  expect_snapshot(summarize_EAF_values(multi_qsip))
 })
 
 test_that("fails on list of non-qsip objects", {